@@ -6,14 +6,14 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
+    "lint": "eslint \"src/**/*.ts\" --fix",
     "prettier": "prettier --write \"src/**/*.{ts,tsx}\"",
     "prepare": "husky"
   },
   "dependencies": {
-    "@tailwindcss/postcss": "^4.1.14",
-    "axios": "^1.12.2",
-    "next": "^15.5.4",
+    "@tailwindcss/postcss": "^4.1.16",
+    "axios": "^1.13.1",
+    "next": "^16.0.1",
     "postcss": "^8.5.6",
     "react": "19.2.0",
     "react-dom": "19.2.0",
@@ -21,27 +21,27 @@
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3.3.1",
-    "@eslint/js": "^9.37.0",
+    "@eslint/js": "^9.39.0",
     "@types/eslint-plugin-security": "^3",
     "@types/node": "^24",
     "@types/react": "^19",
     "@types/react-dom": "^19",
     "@types/uuid": "^11.0.0",
     "autoprefixer": "^10",
-    "daisyui": "^5.1.29",
-    "eslint": "^9.37.0",
-    "eslint-config-next": "^15.5.4",
+    "daisyui": "^5.4.2",
+    "eslint": "^9.39.0",
+    "eslint-config-next": "^16.0.1",
     "eslint-config-prettier": "10.1.8",
     "eslint-plugin-prettier": "5.5.4",
     "eslint-plugin-security": "^3.0.1",
     "eslint-plugin-simple-import-sort": "^12.1.1",
-    "globals": "^16.4.0",
+    "globals": "^16.5.0",
     "husky": "^9.1.7",
     "jest": "^30.2.0",
     "prettier": "3.6.2",
-    "tailwindcss": "^4.1.14",
-    "ts-jest": "29.4.4",
+    "tailwindcss": "^4.1.16",
+    "ts-jest": "29.4.5",
     "typescript": "5.9.3",
-    "typescript-eslint": "8.46.0"
+    "typescript-eslint": "8.46.2"
   }
 }
