@@ -1,5 +1,7 @@
 /** @type {import('next').NextConfig} */
 
-const nextConfig = {};
+const nextConfig = {
+    cacheComponents: true,
+};
 
-module.exports = nextConfig
+export default nextConfig;
